requirement.txt
llama-index
llama-parse
langchain
langchain-openai
qdrant-client
ragas
python-dotenv